update download icon
