+++ src/components/AudioPlayer/PlaybackControls.tsx
@@ -1,20 +1,18 @@
 import { useSelector } from 'react-redux';
-import { selectAudioFile, selectIsPlaying } from 'src/redux/slices/AudioPlayer/state';
+import { selectIsPlaying } from 'src/redux/slices/AudioPlayer/state';
 import { withStopPropagation } from 'src/utils/event';
-import download from 'src/utils/download';
 import styles from './PlaybackControls.module.scss';
 import PlayIcon from '../../../public/icons/play-arrow.svg';
 import ForwardIcon from '../../../public/icons/forward_10.svg';
 import PauseIcon from '../../../public/icons/pause.svg';
 import ReplayIcon from '../../../public/icons/replay_10.svg';
-import DownloadIcon from '../../../public/icons/download.svg';
 import RepeatIcon from '../../../public/icons/ic_repeat_24px 1.svg';
 import Button, { ButtonShape, ButtonSize, ButtonVariant } from '../dls/Button/Button';
 import { triggerPauseAudio, triggerPlayAudio, triggerSeek } from './EventTriggers';
+import DownloadAudioButton from './DownloadAudioButton';
 
 const PlaybackControls = () => {
   const isPlaying = useSelector(selectIsPlaying);
-  const audioFile = useSelector(selectAudioFile);
   return (
     <div className={styles.container}>
       <Button variant={ButtonVariant.Ghost} shape={ButtonShape.Circle} size={ButtonSize.Large}>
@@ -55,14 +53,7 @@ const PlaybackControls = () => {
       >
         <ForwardIcon />
       </Button>
-      <Button
-        onClick={withStopPropagation(() => download(audioFile.audioUrl))}
-        variant={ButtonVariant.Ghost}
-        shape={ButtonShape.Circle}
-        size={ButtonSize.Large}
-      >
-        <DownloadIcon />
-      </Button>
+      <DownloadAudioButton />
     </div>
   );
 };
